(split from avrecode) Add init_cabac_decoder hook to libavcodec and split stream.
This diff threads a hook configuration from the AVCodecContext to the
CABACContext and calls the hook when the decoder finds a CABAC encoded
bit block.

Refactored the hooks into a compressor class. A similar set of hooks
will be used for decompressing.

Use memmem to find the cabac block within the input. A few blocks are
not findable because of escape sequences (can be fixed later on if
needed).

Add a protobuf to store the compressed output. Split the input into two
streams of blocks: raw blocks, and CABAC blocks. The input can be
recovered by interleaving the two lists.