(split from avrecode) Encode CABAC blocks that mostly match the input.
Careful coding around the CABAC trailing bits addresses most of the
mismatch issues. There are still mismatches early in some blocks for
unknown reasons.

Added a simple arithmetic coder in the compressor to measure the benefit
of higher fidelity probability estimates with the same contexts as
CABAC. I was able to save about 0.5% without much tuning (just the decay
rate).